Fix overrun of Field_Tiles array
Random tile search used rand_range(0, w*h) to pick a random tile, but since rand() is inclusive of RAND_MAX, this can result in the invalid index of w*h. Since it's expected, particularly in the case of rand_rangef(0.f, 1.f) that both endpoints are included, the rand_range is now (0, w*h - 1).